requirements.txt
pandas>=1.5.0
numpy>=1.21.0
ta>=0.10.0
python-binance>=1.0.16
yfinance>=0.2.18
streamlit>=1.24.0
openai>=1.0.0
python-dotenv>=0.19.0
pyyaml>=6.0.1
requests>=2.31.0
plotly>=5.18.0
setuptools>=68.0.0 
alpha-vantage==2.3.1